# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenDeX Backend API",
    "version": "0.1.0",
    "summary": "Local control API of the OpenDeX backend (Android multi-window mirroring).",
...
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "contact": {
      "name": "OpenDeX on GitHub",
      "url": "https://github.com/Lordof21/OpenDex"
    }
  },
  "paths": {
...
    "/api/v1/windows": {
      "get": {
        "tags": [
          "windows"
        ],
        "summary": "Get Windows",
        "description": "Lists all active window sessions.",
        "operationId": "get_windows_api_v1_windows_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WindowState"
                  },
                  "type": "array",
                  "title": "Response Get Windows Api V1 Windows Get"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          }
        }
      }
    },
    "/api/v1/windows/open": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Open Window",
        "description": "Spawns an isolated scrcpy session on a virtual display for an app.",
        "operationId": "open_window_api_v1_windows_open_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWindowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindowHandle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/windows/mode": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Set Display Mode",
        "description": "Updates display mode (maximized or windowed).",
        "operationId": "set_display_mode_api_v1_windows_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/windows/close": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Close Window",
        "description": "Closes an active window session and frees on-device encoder resources.",
        "operationId": "close_window_api_v1_windows_close_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWindowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/windows/handoff": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Handoff Window",
        "description": "Gracefully handoffs an OpenDeX window to the physical phone (Stealth DPI —\ndensity equalization; see HandoffManager.handoff_to_phone).",
        "operationId": "handoff_window_api_v1_windows_handoff_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffWindowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/windows/reclaim": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Reclaim Window",
        "description": "Brings an app that moved to Display 0 (the phone) back to the PC window\n(Stealth DPI — see handoff_window).",
        "operationId": "reclaim_window_api_v1_windows_reclaim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReclaimWindowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/windows/applock/retry": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Retry App Lock",
        "description": "Re-runs the AppLock unlock-wait sequence for a window still stuck\nbehind an OEM lock screen (AppLockOverlay's \"Tekrar Dene\" button) —\navoids forcing the user to close and cold-relaunch the app after a\nfailed fingerprint/PIN attempt.",
        "operationId": "retry_app_lock_api_v1_windows_applock_retry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryAppLockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/windows/visibility": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Set Visibility",
        "description": "Sets window visibility state (visible, minimized, occluded).",
        "operationId": "set_visibility_api_v1_windows_visibility_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/windows/focus": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Focus Window",
        "description": "Bumps window z-index to top and updates MRU.",
        "operationId": "focus_window_api_v1_windows_focus_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/windows/resize": {
      "post": {
        "tags": [
          "windows"
        ],
        "summary": "Resize Window",
        "description": "Dynamically reconfigures resolution and DPI of a running window's virtual display.",
        "operationId": "resize_window_api_v1_windows_resize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindowHandle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
...
    "/api/v1/window/{window_id}/xml-elements": {
      "get": {
        "tags": [
          "windows"
        ],
        "summary": "Get Window Xml Elements",
        "description": "Zero-Apps ADB XML Element Inspector (Faz 3+ Deneysel Mod).",
        "operationId": "get_window_xml_elements_api_v1_window__window_id__xml_elements_get",
        "parameters": [
          {
            "name": "window_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Window Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The per-user API token (`OPENDEX_API_TOKEN` or `~/.opendex/api-token`)."
      },
      "tokenQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "The same token as a query parameter — accepted on GET requests and WebSocket handshakes only."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "The API token is missing or wrong (WebSocket: close code 4401).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "ForeignOrigin": {
        "description": "The request carries an `Origin` that is not one of the app's own origins, or a `Host` that is not a loopback name (WebSocket: close code 4403).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8710",
      "description": "Default local listener (OPENDEX_HTTP_HOST / OPENDEX_HTTP_PORT)"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "tokenQuery": []
    }
  ]
}